Workaround that resolves issue with CouchDB 1.4.0 returning empty documents on /_users/_all_docs
fix for https://github.com/hoodiehq/hoodie-server/issues/140
which is itself caused by https://issues.apache.org/jira/browse/COUCHDB-1888

This may not be the right place for doing this, but it seemed logical given that it deals with account management, and the change event would receive empty documents otherwise.

Also, the list of fields is hard coded. I thought a dynamic approach to specifying the fields to make public was the right approach, but doing it without accidentally exposing secure information that a future plugin  might introduce to the user documents is going to be a tricky thing to do.